FILE: moves Options class into game folder
