+++ libs/spirit-islander/util/src/lib/game/options.ts
@@ -1,10 +1,10 @@
-import { ADVERSARIES, Adversary, AdversaryLevelId } from './game/adversaries';
-import { BOARDS, BalancedBoardName, Board } from './game/boards';
-import { ExpansionName } from './game/expansions';
-import { MAPS, Map, MapName } from './game/maps';
-import { SCENARIOS, Scenario, ScenarioName } from './game/scenarios';
-import { SPIRITS, Spirit, SpiritName } from './game/spirits';
-import { getOptionsByExpansion } from './game/get-options-by-expansion';
+import { ADVERSARIES, Adversary, AdversaryLevelId } from './adversaries';
+import { BOARDS, BalancedBoardName, Board } from './boards';
+import { ExpansionName } from './expansions';
+import { MAPS, Map, MapName } from './maps';
+import { SCENARIOS, Scenario, ScenarioName } from './scenarios';
+import { SPIRITS, Spirit, SpiritName } from './spirits';
+import { getOptionsByExpansion } from './get-options-by-expansion';
 
 export class Options {
   private _spirits = SPIRITS;
